compile
svn path=/trunk/kdenetwork/kopete/; revision=212680

+++ libkopete/kopetemessage.cpp
@@ -145,7 +145,7 @@ void KopeteMessage::compareColors( QColor &colorFg, QColor &colorBg )
 	colorBg.hsv( &h2, &s2, &v2 );
 	vDiff = v1 - v2;
 
-	if( h1 == s1 && h2 == s2 && ( abs(vDiff) <= 150 ) )
+	if( h1 == s1 && h2 == s2 && ( abs((double)vDiff) <= 150 ) )
 		colorFg = QColor( h2, s2, (v1 + 127) % 255, QColor::Hsv );
 }
